UI调整

+++ src/main/java/com/taisheng/now/bussiness/bean/result/market/DingdanxiangqingGoodBean.java
@@ -1,10 +1,12 @@
 package com.taisheng.now.bussiness.bean.result.market;
 
+import java.math.BigDecimal;
+
 public class DingdanxiangqingGoodBean {
     public String goodsId;
     public String orderGoodsId;
     public String picUrl;
     public String name;
-    public int price;
+    public BigDecimal price;
     public int number;
 }
