+++ src/main/java/com/taisheng/now/bussiness/bean/result/market/DingdanxiangqingResultBean.java
@@ -1,5 +1,6 @@
 package com.taisheng.now.bussiness.bean.result.market;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class DingdanxiangqingResultBean {
@@ -8,7 +9,7 @@ public class DingdanxiangqingResultBean {
         public String orderId;
         public String orderSn;
         public int goodsNumber;
-        public int totalPrice;
+        public BigDecimal totalPrice;
         public String status;
         public List<DingdanxiangqingGoodBean> list;
         public String message;
